+++ include/learning/tree/ot.h
@@ -29,18 +29,18 @@
 
 class ObliviousRT : public RegressionTree {
  public:
-  ObliviousRT(unsigned int nodes, quickrank::data::Dataset *dps, double *labels,
-              unsigned int minls, unsigned int treedepth)
+  ObliviousRT(size_t nodes, quickrank::data::Dataset *dps, double *labels,
+              size_t minls, size_t treedepth)
       : RegressionTree(nodes, dps, labels, minls),
         treedepth(treedepth) {
   }
   void fit(RTNodeHistogram *hist);
 
  protected:
-  const unsigned int treedepth = 0;
+  const size_t treedepth = 0;
 
  private:
-  void fill(double **sumvar, const unsigned int nfeaturesamples,
+  void fill(double **sumvar, const size_t nfeaturesamples,
             RTNodeHistogram const *hist);
   const double invalid = -DBL_MAX;
 };
